@@ -8,8 +8,7 @@ import {
   InjectionToken
 } from '@angular/core'
 
-import { Observable } from 'rxjs/Observable'
-import { BehaviorSubject } from 'rxjs/BehaviorSubject'
+import { Observable, BehaviorSubject } from 'rxjs'
 import 'rxjs/add/operator/switchMap'
 
 import { PopinContainerComponent } from './popin-container.component'
